+++ src/aoc5.rs
@@ -7,13 +7,13 @@ pub fn run() {
     println!("5:2 {}", run_2(&input));
 }
 
-fn run_1(input: &str) -> i32 {
+fn run_1(input: &str) -> i64 {
     let mut cpu = super::intcode::CPU::new(input);
     cpu.run(&mut vec![1]);
     cpu.output[cpu.output.len() - 1]
 }
 
-fn run_2(input: &str) -> i32 {
+fn run_2(input: &str) -> i64 {
     let mut cpu = super::intcode::CPU::new(input);
     cpu.run(&mut vec![5]);
     cpu.output[cpu.output.len() - 1]
